Add files via upload
This is the first version of my code. Most of it was generated by AI, but this version includes some of my own edits. Changes I made to the code are that I made the reset time 00:00 and that the minute time increases by increments of 5. Another change I made was to make it so that each time the alarm time is changed, the SSD will show the time it is set at for a few seconds before switching back to the current time.