Update and rename Version_1_coffee_code.ino to coffee_maker_code.ino
This is my second version of code in which I added the limit switch to my program. I did this by adding the variable alarmOn, which was true when the current time equaled the alarm time. This allowed me to simply make a condition to turn on the buzzer when the alarmOn was true and the limit switch was being pressed.