Update coffee_maker_code.ino
This is the final version of my code. The only change I made is adding the pump to my code, so that it mirrored the behavior of my buzzer, turning on when alarmOn was true and the limit switch was pressed, and off when alarmOn was false and the limit switch was not being pressed.